requirements.txt
djangorestframework
python-decouple
psycopg2-binary
celery
redis
pytelegrambotapi
pyqt5
docker
pytest
djangorestframework-simplejwt
